pyproject.toml
[project]
name = "biomarker-ai"
version = "0.1.0"
description = "CLI tool for AI-assisted biomarker analysis"
authors = [{name = "AI Developer"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2",
    "openpyxl>=3.1",
    "typer>=0.12",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "requests>=2.31",
    "pydantic>=2.8"
]

[project.scripts]
biomarker-ai = "biomarker_ai.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["biomarker_ai"]

[tool.setuptools.package-data]
"biomarker_ai" = ["py.typed"]
